+++ CHANGELOG.md
@@ -7,10 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-## [0.2.8] - 
+## [0.2.8] - 2026-01-14
 
 ### Added
-- 
+- [Ticket 9](https://github.com/Xenon1/NextUp/issues/9) - Added custom popups that only ask for confirmation if destructive
 
 ### Changed
 - [Ticket 8](https://github.com/Xenon1/NextUp/issues/8) - Altered text in in show and anime details from **Currently Watching** to **Last Episode Watched**